requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.14.0
peft>=0.4.0
bitsandbytes>=0.41.0

# 语音处理
openai-whisper
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11

# 中文处理
jieba>=0.42.1

# Web框架
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.1.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 其他工具
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
